Refactor: db 부분 리팩토링
- 서비스 디렉토리 추가
- 구조 살짝 변경
