+++ server/src/controller/document.js
@@ -1,9 +1,9 @@
-const db = require("../api/db");
+const documentService = require("../services/document");
 
 const save = async (req, res) => {
   // const userId = req.params.userId
   const { userId, docContent } = req.body;
-  const result = await db.document.save(userId, docContent);
+  const result = await documentService.save(userId, docContent);
   if (result) {
     res.status(200).send();
   } else {
@@ -14,7 +14,7 @@ const save = async (req, res) => {
 const load = async (req, res) => {
   // const userId = req.params.userId
   const userId = "boost";
-  const result = await db.document.load(userId);
+  const result = await documentService.load(userId);
   if (result) {
     res.status(200).send(result[0].content);
   } else {
